% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{add_coverage}
\alias{add_coverage}
\title{Coverage management functions}
\usage{
add_coverage(coverage, name, overwrite = FALSE)
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{name}{A \code{character} text with coverage name.}

\item{overwrite}{A \code{logical} value indicating if coverage can be overwritten. Default \code{FALSE}.}
}
\value{
\code{TRUE} if coverage is registered.
}
\description{
Register a coverage object in the local coverage list.
}
