% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{add_remote}
\alias{add_remote}
\title{Coverage management functions}
\usage{
add_remote(name, location, bucket = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{name}{A \code{character} text with the remote name.}

\item{location}{A \code{character} text with a file path or url to a YAML manifest list file.}

\item{bucket}{An optional \code{character} text with bucket AmazonS3 name. Default \code{NULL}.}

\item{overwrite}{A \code{logical} value indicating if remote can be overwritten. Default \code{FALSE}.}

\item{...}{Any extra arguments to be passed to \code{aws.s3} package to access the bucket.}
}
\value{
\code{TRUE} if a remote is added.
}
\description{
Add a new remote to the local remotes list.
}
