% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{apply_bricks}
\alias{apply_bricks}
\title{Coverage functions}
\usage{
apply_bricks(coverage, fun, ...)
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{fun}{A function to be called for each brick that receives the bricks' attributes.}

\item{...}{A set of \code{names} or \code{character} vector with the bands to be returned.}
}
\value{
A rasters \code{data.frame} object.
}
\description{
Apply a function on each \code{coverage}'s bricks object. Parameter
\code{fun} is a function with bricks attributes names as parameters. A helper
function \code{func_brick} can be used to create a prototype of this function easily.
}
