% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{apply_bricks_cluster}
\alias{apply_bricks_cluster}
\title{Coverage functions}
\usage{
apply_bricks_cluster(coverage, fun, ..., clusters = 1,
  cluster_type = c("PSOCK", "FORK"))
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{fun}{A function to be called for each brick that receives the bricks' attributes.}

\item{...}{A set of \code{names} or \code{character} vector with the bands to be returned.}

\item{clusters}{An \code{integer} with the number of clusters to be created.}

\item{cluster_type}{A \code{character} text with the type of cluster to be created.}
}
\value{
A rasters \code{data.frame} object.
}
\description{
Apply a function on each \code{coverage}'s bricks object. Parameter
\code{fun} is a function with bricks attributes names as parameters. A helper
function \code{func_brick} can be used to create a prototype of this function easily.
The function creates processing clusters using package \code{parallel}.
}
