% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks_df.R
\name{bricks}
\alias{bricks}
\title{Bricks \code{data.frame} function}
\usage{
bricks(files, timeline, bands = coverage::MOD13Q1_bands)
}
\arguments{
\item{files}{A \code{character} vector listing all raster files location to be used as bricks.}

\item{timeline}{A \code{date} vector indicating all dates of each brick time series.}

\item{bands}{A \code{data.frame} containing 'name', 'short_name', 'min', 'max',
'fill', and 'scale' columns for each band to be retrieved from \code{files}.}
}
\value{
A bricks \code{data.frame} object.
}
\description{
Creates a new \code{bricks} object used to make coverages.
A \code{bricks} object contains four mandatory attributes: 'file', 'key', 'band', and 'timeline'.
The 'key' and 'band' attributes are extracted from file name.
A bricks 'key' is a name that identifies all bands files of one brick. The bricks keys
text are trimmed on the following default characters \code{getOption("trim_keys", "_. ")}.
A bricks 'band' is the name of the band.
}
