% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{check_coverage}
\alias{check_coverage}
\title{Coverage functions}
\usage{
check_coverage(coverage, touch_files = TRUE)
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{touch_files}{Verify if raster files exists.}
}
\value{
The \code{coverage} object with \code{class} attribute if pass in all checks.
}
\description{
Checks the coverage validity.
}
