% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_functions}
\alias{config_functions}
\alias{.default_config}
\alias{save_config}
\alias{load_config}
\alias{link_repository}
\alias{unlink_repository}
\alias{list_repositories}
\title{Config functions}
\usage{
.default_config()

save_config()

load_config()

link_repository(name, location)

unlink_repository(name)

list_repositories(prefix = NULL)
}
\arguments{
\item{name}{A \code{character} text with repository or cube entry name.}

\item{location}{An \code{character} string with a valid json location.}

\item{prefix}{A \code{character} containing an prefix entry name
to be filtered.}
}
\value{
A \code{list} data structure.

None.

None.

A \code{repository} object.

None.

A \code{list} of repositories.
}
\description{
These functions provides a basic management of repositories.
A repository is an entry point to a cube provider that can maintain many cubes.
}
\details{
The function \code{link_repository} fetches the repository definition and add an
entry into the package config using \code{name} as key.

The function \code{rm_repository} remove the repository entry from the package config.

The function \code{list_repositories} lists all registered repository
entries. Use \code{prefix} parameter to filter the entries by name.
}
\section{Functions}{
\itemize{
\item \code{.default_config}: Return the default config.

\item \code{save_config}: Save the current config state.

\item \code{load_config}: Load the current config state.

\item \code{link_repository}: Adds a repository definition to the package config.

\item \code{unlink_repository}: Remove a repository entry from the package config.

\item \code{list_repositories}: Lists all registered repositories.
}}

\examples{
list_repositories()
x <- repository("localhost")
repository_name(x)   # shows 'localhost'
list_cubes(x)   # list cubes in 'localhost'

}
\seealso{
\code{\link{list_cubes}}
}
