% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{coverage}
\alias{coverage}
\title{Coverage functions}
\usage{
coverage(bricks, manifest = manifest())
}
\arguments{
\item{bricks}{A bricks \code{data.frame}.}

\item{manifest}{A description \code{list} returned by \code{coverage.manifest} function.}
}
\value{
A \code{coverage} object.
}
\description{
Create a new \code{coverage} object. A \code{coverage} object
contains all information about the coverage authoring, data policy, and
bricks definitions.
}
