% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cube_bands}
\alias{cube_bands}
\title{Cubes management functions}
\usage{
cube_bands(cube)
}
\arguments{
\item{cube}{An \code{EOCubes_cube} data structure.}
}
\value{
A \code{list} with all registered bands in the given cube.
}
\description{
Lists the registered bands of a cube.
}
