% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cube_crs}
\alias{cube_crs}
\title{Cubes management functions}
\usage{
cube_crs(cube)
}
\arguments{
\item{cube}{An \code{EOCubes_cube} data structure.}
}
\value{
A CRS \code{character} string of a given cube.
}
\description{
Get the registered coordinate reference system of a cube.
}
