% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cube_functions}
\alias{cube_functions}
\alias{cube}
\alias{cube_name}
\alias{cube_bands}
\alias{cube_crs}
\title{Cube functions}
\usage{
cube(name, remote = default_remote())

cube_name(cube)

cube_bands(cube)

cube_crs(cube)
}
\arguments{
\item{name}{A \code{character} text with cube name.}

\item{remote}{A \code{character} text with remote name.}

\item{cube}{An \code{EOCubes_cube} object.}
}
\value{
An \code{EOCubes_cube} object.

A \code{character} string.

An \code{EOCubes_bandlist} object.

A \code{character} string.
}
\description{
These functions provides the basic operations over a cube
(\code{EOCubes_cube}) object.
}
\section{Functions}{
\itemize{
\item \code{cube}: Fetches a cube from remote.

\item \code{cube_name}: Returns the name of a cube.

\item \code{cube_bands}: Lists the all registered bands in a cube.

\item \code{cube_crs}: Get the registered coordinate reference system (CRS) of a cube.
}}

\examples{
x <- remote("localhost")
cub1 <- cube("MOD13Q1/006", x)
cube_name(cub1)   # show the entry name 'MOD13Q1/006'
cube_bands(cub1)  # show bands and its meta data
cube_crs(cub1)    # show crs string

}
\seealso{
\code{\link{remote}}
}
