% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cube_tiles}
\alias{cube_tiles}
\title{Cubes management functions}
\usage{
cube_tiles(cube, which = NULL, prefix = NULL)
}
\arguments{
\item{cube}{A \code{EOCubes_cube} data structure.}

\item{which}{A \code{logical} or \code{integer} vector indicating which tile to retrieve.}

\item{prefix}{A \code{character} vector containing tile name prefix to be filtered.}
}
\value{
A \code{list} of tiles in a cube or \code{NULL} if filter does not return tiles.
}
\description{
Lists all registered tiles in a cube.
}
\details{
The \code{which} vector must have the same length as cube tiles list. It can
be used with \code{sf} dependent function \code{tiles_intersects} to retrieve
only a subset of tiles that intersects some geometry.
}
