% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{describe_coverage}
\alias{describe_coverage}
\title{Coverage management functions}
\usage{
describe_coverage(name)
}
\arguments{
\item{name}{A \code{character} with the coverage name.}
}
\value{
A coverage manifest \code{list}. If the informed name does not
correspond to any coverage, return \code{NULL}.
}
\description{
Describes a given coverage by name. The name is searched first in
the local manifest list file. If not found, then it is searched in remote
manifests file. The function returns the manifest of the corresponing coverage name.
}
