% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{.check_bands}
\alias{.check_bands}
\title{Internal coverage functions}
\usage{
.check_bands(bands)
}
\arguments{
\item{bands}{A bands definition \code{list} object.}
}
\value{
\code{TRUE} if pass in all checks.
}
\description{
Check for validity of a coverages bands definition.
}
