% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{.check_bricks}
\alias{.check_bricks}
\title{Internal coverage functions}
\usage{
.check_bricks(bricks)
}
\arguments{
\item{bricks}{A bricks \code{list} object.}
}
\value{
\code{TRUE} if pass in all checks.
}
\description{
Check for validity of a coverages bricks list.
}
