% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks_df.R
\name{.check_bricks_df}
\alias{.check_bricks_df}
\title{Internal bricks \code{data.frame} functions}
\usage{
.check_bricks_df(bricks)
}
\arguments{
\item{bricks}{A bricks \code{data.frame}.}
}
\value{
\code{TRUE} if pass in all checks.
}
\description{
Check if the bricks \code{data.frame} attribute values are consistent and
all files are reachable.
A progress bar will be showed if the number of files in bricks is greater or equal
the value of \code{getOption("progress_bar", 10)}.
}
