% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{.check_coverage}
\alias{.check_coverage}
\title{Internal coverage functions}
\usage{
.check_coverage(coverage)
}
\arguments{
\item{coverage}{A \code{coverage} object.}
}
\value{
\code{TRUE} if pass in all checks.
}
\description{
Checks the coverage validity.
}
