% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{.check_coverage_timeline}
\alias{.check_coverage_timeline}
\title{Internal coverage functions}
\usage{
.check_coverage_timeline(coverage)
}
\arguments{
\item{coverage}{A \code{coverage} object.}
}
\value{
\code{TRUE} if pass in all checks.
}
\description{
Check for validity of a coverage timeline field.
}
