% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{.check_manifest}
\alias{.check_manifest}
\title{Internal manifest functions}
\usage{
.check_manifest(manifest)
}
\arguments{
\item{manifest}{A manifest \code{list} object.}
}
\value{
\code{TRUE} if pass in all checks.
}
\description{
Check for validity of a coverages manifest list.
}
