% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{.check_timeline}
\alias{.check_timeline}
\title{Internal coverage functions}
\usage{
.check_timeline(timeline, bricks_df)
}
\arguments{
\item{timeline}{A \code{vector} sequence with dates.}

\item{bricks_df}{A bricks \code{data.frame} object.}
}
\value{
\code{TRUE} if pass in all checks.
}
\description{
Check for validity of a coverage timeline field.
}
