% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{.fetch_tiles}
\alias{.fetch_tiles}
\title{Cubes management functions}
\usage{
.fetch_tiles(tile_lst)
}
\arguments{
\item{tile_lst}{A \code{EOCube_tilelist} data structure.}
}
\value{
A \code{list} of tiles contents in a cube.
}
\description{
Fetch all registered tiles in a cube.
}
