% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.file.metadata}
\alias{.file.metadata}
\title{Internal functions}
\usage{
.file.metadata(file)
}
\arguments{
\item{file}{A \code{character} text indicating a file path (local or in the web) to be opened.}
}
\value{
A \code{data.frame} object containing the metadata.
}
\description{
Retrieve rater file geographycal metadata information.
}
