% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.file.open}
\alias{.file.open}
\title{Internal functions}
\usage{
.file.open(file)
}
\arguments{
\item{file}{A \code{character} text indicating a file path (local or in the web) to be opened.}
}
\value{
A \code{RasterBrick} object.
}
\description{
Open a brick file using \code{raster} package. If \code{file} is a url, it uses
GDAL driver \code{vsicurl}.
}
