% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.files.bands}
\alias{.files.bands}
\title{Internal functions}
\usage{
.files.bands(files, bands)
}
\arguments{
\item{files}{A \code{character} vector with all files paths.}

\item{bands}{A \code{bands} object \code{data.frame} containing all bands definitions.}
}
\value{
A \code{character} vector with all bands corresponding to the input file names.
}
\description{
Extracts from file names the name of the bands according to a bands definition.
}
