% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.files.keys}
\alias{.files.keys}
\title{Internal functions}
\usage{
.files.keys(files, bands)
}
\arguments{
\item{files}{A \code{character} vector with all files paths.}

\item{bands}{A \code{bands} object \code{data.frame} containing all bands definitions.}
}
\value{
A \code{character} vector with all keys corresponding to the input file names.
}
\description{
Extracts from file names the bricks 'key' values. The 'key' groups
files of different bands into one bricks definition. According to
\code{getOption("trim_keys", "_. ")} option, the leading and trainling characters
will be trimmed of the key value.
}
