% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.files.metadata}
\alias{.files.metadata}
\title{Internal functions}
\usage{
.files.metadata(files)
}
\arguments{
\item{files}{A \code{character} vector with all files paths.}
}
\value{
A metadata \code{data.frame} with all metadata values.
}
\description{
Retrieve some metadata from the bricks' raster files.
A progress bar will be showed if the number of files in bricks is greater or equal
the value of \code{getOption("progress_bar", 10)}.
}
