% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.filter_prefix}
\alias{.filter_prefix}
\title{Internal functions}
\usage{
.filter_prefix(prefix, named_list)
}
\arguments{
\item{prefix}{A \code{character} containing a name prefix to be filtered.}

\item{named_list}{Any named \code{list} containing the data to be filtered.}
}
\value{
A \code{list} object containing the read data.
}
\description{
Returns a subset of a given named list.
}
