% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.is_empty.character}
\alias{.is_empty.character}
\title{Internal functions}
\usage{
\method{.is_empty}{character}(x)
}
\arguments{
\item{x}{A \code{character} vector to be tested.}
}
\value{
\code{TRUE} if \code{x} has length zero, has at least one \code{NA}, or \code{""} value. \code{FALSE} otherwise.
}
\description{
Internal functions
}
