% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.is_empty.data.frame}
\alias{.is_empty.data.frame}
\title{Internal functions}
\usage{
\method{.is_empty}{data.frame}(x)
}
\arguments{
\item{x}{A \code{data.frame} object to be tested.}
}
\value{
\code{TRUE} if \code{data.frame} has zero columns or zero rows. \code{FALSE} otherwise.
}
\description{
Internal functions
}
