% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.is_empty.default}
\alias{.is_empty.default}
\title{Internal functions}
\usage{
\method{.is_empty}{default}(x)
}
\arguments{
\item{x}{Any object to be tested.}
}
\value{
\code{TRUE} if \code{x} is \code{NULL}. \code{FALSE} otherwise.
}
\description{
Internal functions
}
