% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.is_empty.list}
\alias{.is_empty.list}
\title{Internal functions}
\usage{
\method{.is_empty}{list}(x)
}
\arguments{
\item{x}{A \code{list} to be tested.}
}
\value{
\code{TRUE} if \code{x} has length zero. \code{FALSE} otherwise.
}
\description{
Internal functions
}
