% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.is_empty.numeric}
\alias{.is_empty.numeric}
\title{Internal functions}
\usage{
\method{.is_empty}{numeric}(x)
}
\arguments{
\item{x}{A \code{numeric} vector to be tested.}
}
\value{
\code{TRUE} if \code{x} has length zero or has at least one \code{NA}. \code{FALSE} otherwise.
}
\description{
Internal functions
}
