% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.is_empty}
\alias{.is_empty}
\title{Internal functions}
\usage{
.is_empty(x)
}
\arguments{
\item{x}{Any object to be tested.}
}
\value{
\code{TRUE} if \code{x} is empty. \code{FALSE} otherwise.
}
\description{
Checks if a value is empty. The meaning of empty state depends on \code{x} class.
If \code{x} is a \code{data.frame}, the object is empty if it has zero columns or zero rows.
If \code{x} is a \code{character} vector, it is empty if \code{x} has length zero, has at least one \code{NA},
or \code{""} value. If \code{x} is a \code{numeric} vector, it is empty if it has length zero or has
at least one \code{NA} as its element. If \code{x} is a \code{list}, it is empty if it has length zero.
Otherwise, it is empty if it is \code{NULL} or has length zero.
}
