% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.open_json}
\alias{.open_json}
\title{Internal functions}
\usage{
.open_json(location, cache)
}
\arguments{
\item{location}{A \code{character} text indicating a file path (local or in the web) to be opened.}

\item{cache}{A \code{logical} indicating if result can be retrieved from cache.}
}
\value{
A \code{list} object containing the read data.
}
\description{
Open a json file and return its structure as a \code{list}.
}
