% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.open_multiple_jsons}
\alias{.open_multiple_jsons}
\title{Internal functions}
\usage{
.open_multiple_jsons(locations, cache, progress_bar)
}
\arguments{
\item{locations}{A \code{character} text indicating a file path (local or in the web) to be opened.}

\item{cache}{A \code{logical} indicating if result can be retrieved from cache.}

\item{progress_bar}{A \code{logical} indicating if a progress bar must be created.}
}
\value{
A \code{list} object containing the read data.
}
\description{
Open a json file and return its structure as a \code{list}.
}
