% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3.R
\name{.s3_url}
\alias{.s3_url}
\title{Internal AWS S3 functions}
\usage{
.s3_url(bucket, object)
}
\arguments{
\item{bucket}{A \code{character} text informing the remote name where coverage will be published.}

\item{object}{A \code{character} text with the object key name in bucket.}
}
\value{
A \code{character} vector of object's url.
}
\description{
This function returns an url to an bucket's object in the Amazon S3 service.
The default base url text used to form the final url is
given by \code{getOption("s3_base_url", "https://\%s.s3.amazonaws.com/\%s")}, where
the two formating text parameters \code{\%s} correspond to the bucket name and
the object key, respectively.
}
