% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.save_json}
\alias{.save_json}
\title{Internal functions}
\usage{
.save_json(x, file)
}
\arguments{
\item{x}{A \code{list} data to be saved.}

\item{file}{A \code{character} text indicating a file path (local) to be saved.}
}
\value{
A \code{data.frame} object containing the metadata.
}
\description{
Save a \code{list} data as a json file.
}
