% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.select_prefix}
\alias{.select_prefix}
\title{Internal functions}
\usage{
.select_prefix(prefix, named_list)
}
\arguments{
\item{prefix}{A \code{character} containing a name prefix to be filtered.}

\item{named_list}{Any named \code{list} containing the data to be filtered.}
}
\value{
A \code{logical} vector.
}
\description{
Returns a \code{logical} vector indicating which elements in
named list are selected.
}
