% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{.set_definition.manifest}
\alias{.set_definition.manifest}
\title{Internal manifest functions}
\usage{
.set_definition.manifest(manifest, definition)
}
\arguments{
\item{manifest}{A coverage manifest \code{list}.}

\item{definition}{A \code{character} text with a location of YML definition.}
}
\value{
A \code{list} containing all pairs \code{name = value} of coverage description attributes.
}
\description{
Changes the location of file definition of the coverage.
}
