% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{fetch}
\alias{fetch}
\alias{.fetch}
\title{Cubes management functions}
\usage{
.fetch(data)
}
\arguments{
\item{data}{A \code{list} of resources.}
}
\value{
A \code{list} of resource contents.
}
\description{
Fetch the contents of a list of resources.
}
\details{
Data is a \code{list} of resources returned by functions \code{list_*()}.
}
