% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{fetch_cube}
\alias{fetch_cube}
\alias{.fetch_cube}
\title{Cubes management functions}
\usage{
.fetch_cube(name, remote = default_remote())
}
\arguments{
\item{name}{A \code{character} text with cube name.}

\item{remote}{A \code{character} text with remote name.}
}
\value{
A \code{cube} data structure.
}
\description{
Fetches a registered cube
}
