% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{fetch_layers}
\alias{fetch_layers}
\alias{.fetch_layers}
\title{Cubes management functions}
\usage{
.fetch_layers(items, start_date = NULL, end_date = NULL,
  count = NULL)
}
\arguments{
\item{items}{A \code{list} of any \code{cube*} data structure.}

\item{start_date}{A \code{date} value to filter layers by date.}

\item{end_date}{A \code{date} value to filter layers by date.}

\item{count}{An \code{integer} with the mandatory number of layers for each stack.}
}
\value{
A \code{list} of layers references.
}
\description{
Fetch selected layers from stacks.
}
