% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{fetch_stacks}
\alias{fetch_stacks}
\alias{.fetch_stacks}
\title{Cubes management functions}
\usage{
.fetch_stacks(items, prefix = NULL)
}
\arguments{
\item{items}{A \code{list} of any \code{cube*} data structure.}

\item{prefix}{A \code{character} vector containing stack name prefix(es)}
}
\value{
A \code{list} of stacks contents.
}
\description{
Fetch all registered stacks.
}
