% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{fetch_tiles}
\alias{fetch_tiles}
\alias{.fetch_tiles}
\title{Cubes management functions}
\usage{
.fetch_tiles(items, prefix = NULL)
}
\arguments{
\item{items}{A \code{list} of any \code{cube*} data structure.}

\item{prefix}{A \code{character} vector containing tile name prefix(es)}
}
\value{
A \code{list} of tiles contents in a cube.
}
\description{
Fetch all registered tiles in a cube.
}
