% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{filter_cube}
\alias{filter_cube}
\title{Cubes management functions}
\usage{
filter_cube(name, remote = default_remote(), tiles = with_tiles(),
  stacks = with_stacks(c("ndvi", "evi")),
  layers = with_layers(start_dates = seq(as.Date("2000-09-01"),
  as.Date("2018-09-01"), by = "years"), end_dates =
  seq(as.Date("2001-08-31"), as.Date("2019-08-31"), by = "years"), count =
  23))
}
\arguments{
\item{name}{A \code{character} text with cube name.}

\item{remote}{A \code{character} text with remote name.}

\item{tiles}{A \code{list} data structure with parameters to select tiles from cube.
Used with \code{with_tiles()} function.}

\item{stacks}{A \code{list} data structure with parameters to select stacks from tiles.
Used with \code{with_stacks()} function.}

\item{layers}{A \code{list} data structure with parameters to select layers from stacks.
Used with \code{with_layers()} function.}
}
\value{
A \code{cube} data structure.
}
\description{
Fetches a registered cube and its data.
}
\details{
The \code{start_dates} and \code{end_dates} parameters values can
be a character vector containing dates in format \code{"\%Y-\%m-\%d"}.
}
