% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{filter_tiles}
\alias{filter_tiles}
\title{Cubes management functions}
\usage{
filter_tiles(cube, which = NULL, prefix = NULL)
}
\arguments{
\item{cube}{A \code{EOCubes_cube} data structure.}

\item{which}{A \code{logical} or \code{integer} vector indicating which tile to retrieve.}

\item{prefix}{A \code{character} vector containing tile name prefix to be filtered.}
}
\value{
A \code{EOCubes_fetched, EOCubes_cube} cube data structure.
}
\description{
Returns a new \code{EOCubes_cube} data structure with filtered tiles and
fetches tiles data from remote.
}
\details{
The \code{which} vector must have the same length as cube tiles list. It can
be used with \code{sf} dependent function \code{tiles_intersects} to retrieve
only a subset of tiles that intersects some geometry.

The returned cube object \code{EOCubes_cube} is a fetched cube \code{EOCubes_fetched},
that means that all bands and rasters references are in memory.
}
