% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{fun_brick}
\alias{fun_brick}
\title{Coverage functions}
\usage{
fun_brick(expr, ...)
}
\arguments{
\item{expr}{An R language expression to be evaluated as the function body.}

\item{...}{Any additional arguments to be included in function evaluation.}
}
\value{
The evaluation of \code{expr} expression.
}
\description{
Returns an enclosure function that can be used as argument of
\code{apply_bricks} and \code{apply_bricks_cluster} functions' parameter \code{fun}.
Parameters passed in \code{...} can be viewed in \code{expr} code.
}
