% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{get_bricks}
\alias{get_bricks}
\title{Coverage functions}
\usage{
get_bricks(coverage, ..., nest_by_key = FALSE)
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{...}{A set of \code{names} of all those bands to be filtered.}

\item{nest_by_key}{A \code{logical} indicating if values must be nested by key. Defalt \code{FALSE}}
}
\value{
A bricks \code{data.frame} object.
}
\description{
Joins bricks with bands information from \code{coverage} object.
The values can be nested in lists by each key.
}
