% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{get_coverage}
\alias{get_coverage}
\title{Coverage management functions}
\usage{
get_coverage(name)
}
\arguments{
\item{name}{A \code{character} text with coverage name.}
}
\value{
A \code{coverage} object.
}
\description{
Read a \code{coverage} object from remotes.
}
