% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{get_rasters}
\alias{get_rasters}
\title{Coverage functions}
\usage{
get_rasters(coverage, ...)
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{...}{A set of \code{names} or \code{character} vector with the bands to be returned.}
}
\value{
A \code{list} of rasters \code{data.frame} objects.
}
\description{
Lits rasters from \code{coverage} object.
}
