% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{get_remote}
\alias{get_remote}
\alias{.fetch_remote}
\title{Remotes management functions}
\usage{
.fetch_remote(name)
}
\arguments{
\item{name}{A \code{character} text with remote name.}
}
\value{
A \code{remote} data structure.
}
\description{
Fetches a registered remote.
}
