% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands_definitions.R
\docType{data}
\name{LC8MOD_bands}
\alias{LC8MOD_bands}
\title{Bands definitions}
\format{A \code{data.frame} with 'original_name', 'short_name', 'min', 'max', 'fill', and 'scale'
columns defining all LC8MOD mix bands/attributes metadata.}
\description{
A \code{data.frame} containing bands metadata definitions of Landsat8 SR, MOD13Q1, and MYD13Q1 composite product.
}
