% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{list_coverages}
\alias{list_coverages}
\title{Coverage management functions}
\usage{
list_coverages(local_only = FALSE)
}
\arguments{
\item{local_only}{A \code{logical} informing if only local registered coverage
manifests must be returned.}
}
\value{
a named \code{list} with a brief description of all available coverages.
}
\description{
List all registered and remote coverage manifests.
}
