% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{list_cubes}
\alias{list_cubes}
\title{Cubes management functions}
\usage{
list_cubes(remote = default_remote(), prefix = NULL)
}
\arguments{
\item{remote}{A \code{remote} data structure.}

\item{prefix}{A \code{character} containing cube name prefix to be filtered.}
}
\value{
A \code{list} of cubes.
}
\description{
Lists all registered cubes in a remote.
}
