% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{list_layers}
\alias{list_layers}
\alias{.list_layers}
\title{Cubes management functions}
\usage{
.list_layers(item, start_date = NULL, end_date = NULL)
}
\arguments{
\item{item}{Any \code{cube*} data structure.}

\item{start_date}{A \code{date} value to filter layers by date.}

\item{end_date}{A \code{date} value to filter layers by date.}
}
\value{
A \code{list} of layers' urls pointing to raster resources
or \code{NULL} if \code{item} does not list layers.
}
\description{
Lists all registered layers and organizes it by date intervals.
}
\details{
The \code{start_date} and \code{end_date} parameters values can
be a Date or a character containing a date in format \code{"\%Y-\%m-\%d"}.
}
