% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{list_remotes}
\alias{list_remotes}
\title{Remotes management functions}
\usage{
list_remotes(prefix = NULL)
}
\arguments{
\item{prefix}{A \code{character} containing remote name prefix to be filtered.}
}
\value{
A \code{list} of remotes.
}
\description{
Lists all registered remotes.
}
