% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{list_stacks}
\alias{list_stacks}
\alias{.list_stacks}
\title{Cubes management functions}
\usage{
.list_stacks(item, prefix = NULL)
}
\arguments{
\item{item}{Any \code{cube*} data structure.}

\item{prefix}{A \code{character} vector containing stack name prefix(es)}
}
\value{
A \code{list} of stacks or \code{NULL} if \code{item} does not list stacks.
}
\description{
Lists all registered stacks.
}
