% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{list_tiles}
\alias{list_tiles}
\alias{.list_tiles}
\title{Cubes management functions}
\usage{
.list_tiles(item, prefix = NULL)
}
\arguments{
\item{item}{Any \code{cube*} data structure.}

\item{prefix}{A \code{character} vector containing tile name prefix(es)}
}
\value{
A \code{list} of tiles in a cube or \code{NULL} if \code{data} does not list tiles.
}
\description{
Lists all registered tiles in a cube.
}
