% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetched.R
\name{make_stacks}
\alias{make_stacks}
\title{Cubes management functions}
\usage{
make_stacks(cube, bands, start_dates, end_dates, count)
}
\arguments{
\item{cube}{A \code{EOCubes_cube} data structure.}

\item{bands}{A \code{character} vector with band names to be retrieved.}

\item{start_dates}{A \code{date} vector to filter layers by date.}

\item{end_dates}{A \code{date} vector to filter layers by date.}
}
\value{
A \code{EOCubes_organized, EOCubes_cube} cube data structure.
}
\description{
Returns a new \code{EOCubes_cube} data structure with filtered tiles and
fetches tiles data from remote.
}
