% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{management_functions}
\alias{management_functions}
\alias{link_repository}
\alias{unlink_repository}
\alias{link_cube}
\alias{unlink_cube}
\title{Management functions}
\usage{
link_repository(name, location)

unlink_repository(name)

link_cube(location)

unlink_cube(name)
}
\arguments{
\item{name}{A \code{character} text with repository or cube entry name.}

\item{location}{An \code{character} string with a valid json location.}
}
\value{
A \code{repository} object.

None.

A \code{cube} object.

None.
}
\description{
These functions provides a basic management of repositories.
A repository is an entry point to a cube provider that can maintain many cubes.
}
\details{
The function \code{link_repository} fetches the repository definition and add an
entry into the package root using \code{name} as key.

The function \code{rm_repository} remove the repository entry from the package root.

The function \code{add_cube} fetches the repository definition and add an
entry into the package root using \code{name} as key.

The function \code{rm_cube} removes the cube reference from the repository 'localhost'.
}
\section{Functions}{
\itemize{
\item \code{link_repository}: Adds a repository definition to the package root.

\item \code{unlink_repository}: Remove a repository entry from the package root.

\item \code{link_cube}: Add a cube reference to the local repository.

\item \code{unlink_cube}: Remove a cube reference.
}}

\examples{
list_repositories()
x <- repository("localhost")
repository_name(x)   # shows 'localhost'
list_cubes(x)   # list cubes in 'localhost'

}
\seealso{
\code{\link{list_cubes}}
}
