% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest}
\alias{manifest}
\title{Manifest functions}
\usage{
manifest(description = NA_character_, authors = NA_character_,
  references = NA_character_, citation = NA_character_,
  data_policy = NA_character_, maintainer = NA_character_,
  definition = NA_character_)
}
\arguments{
\item{description}{A \code{character} text describing the coverage.}

\item{authors}{A \code{character} vector with coverage authors.}

\item{references}{A \code{character} vector with relevant references of data used.}

\item{citation}{A \code{character} text with coverage citation.}

\item{data_policy}{A \code{character} vector with urls or text informing data policy.}

\item{maintainer}{A \code{character} text with coverage maintainer name and/or contact.}

\item{definition}{A \code{character} text with a location of YML definition.}
}
\value{
A \code{list} containing all pairs \code{name = value} of coverage description attributes.
}
\description{
Generates a list with coverage description attributes.
This list is used to describe the data, the authors, the data policy,
relevant references to data sources used, and the data maintainer.
}
