% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands_definitions.R
\docType{data}
\name{MYD13Q1_bands}
\alias{MYD13Q1_bands}
\title{Bands definitions}
\format{A \code{data.frame} with 'original_name', 'short_name', 'min', 'max', 'fill', and 'scale'
columns defining all MYD13Q1 bands/attributes metadata.}
\description{
A \code{data.frame} containing bands metadata definitions of MYD13Q1 product.
}
