% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverages.R
\name{print.coverage}
\alias{print.coverage}
\title{Coverage functions}
\usage{
\method{print}{coverage}(x, ...)
}
\arguments{
\item{x}{A \code{coverage} object to be printed on screen.}

\item{...}{Other parameters passed by \code{print} generic function.}
}
\description{
Prints a readable summary for \code{coverage} object.
}
