% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{publish_coverage}
\alias{publish_coverage}
\title{Coverage management functions}
\usage{
publish_coverage(coverage, name, remote, overwrite = FALSE, ...)
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{name}{A \code{character} text with coverage name.}

\item{remote}{A \code{character} text with a valid remote server (with 'bucket' defined).}

\item{overwrite}{A \code{logical} indicating if coverage can be overwritten in bucket. Default \code{FALSE}.}

\item{...}{Any extra arguments to be passed to \code{aws.s3} package to access the bucket.}
}
\value{
\code{TRUE} if coverage is published.
}
\description{
Save a \code{coverage} object to a YAML in a remote S3 server
This function uses \code{aws.s3} package to access the remote server.
You must provide appropriate access key to be able to write in the bucket.
}
