% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{read_coverage}
\alias{read_coverage}
\title{Coverage management functions}
\usage{
read_coverage(location)
}
\arguments{
\item{location}{A \code{character} text indicating a file path or url to a valid YAML coverage definition.}
}
\value{
A \code{coverage} object read from YAML file definition.
}
\description{
Read a \code{coverage} object from YAML file.
}
