% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{rm_coverage}
\alias{rm_coverage}
\title{Coverage management functions}
\usage{
rm_coverage(name)
}
\arguments{
\item{name}{A \code{character} text with coverage name.}
}
\value{
\code{TRUE} if coverage is unregistered.
}
\description{
Unregister a coverage object from the local manifest list file.
}
