% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{rm_remote}
\alias{rm_remote}
\title{Coverage management functions}
\usage{
rm_remote(name)
}
\arguments{
\item{name}{A \code{character} text with the remote name to be removed.}
}
\value{
\code{TRUE} if a remote is removed.
}
\description{
Removes a remote from the local remote list.
}
