% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3.R
\name{s3_list}
\alias{s3_list}
\title{AWS S3 functions}
\usage{
s3_list(bucket, regex = ".*\\\\.tif$", ...)
}
\arguments{
\item{bucket}{A \code{character} text informing an existing bucket.}

\item{regex}{A \code{character} with regular expression indicating which files will be included.}

\item{...}{Any extra arguments to be passed to \code{aws.s3} package to access the bucket.}
}
\value{
A \code{character} vector of object's url.
}
\description{
This function retrieves a list of objects from a bucket of the Amazon S3 service and
generate url for each one. The default base url text used to form the final url is
given by \code{getOption("s3_base_url", "https://\%s.s3.amazonaws.com/\%s")}, where
the two formating text parameters \code{\%s} correspond to the bucket name and
the object key, respectively.

Use \code{regex} parameter to filter those objects of your interest.
}
\note{
You must set \code{Sys.setenv("AWS_ACCESS_KEY_ID", <your_key>)} and
\code{Sys.setenv("AWS_SECRET_ACCESS_KEY", <your_access_key>)} to have access to
an aws.s3 \code{bucket} object. If you prefer, you can pass access parameters to
\code{...} argument the same used by \code{aws.s3} package.
For more details see https://github.com/cloudyr/aws.s3.
}
