% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{save_coverage}
\alias{save_coverage}
\title{Coverage management functions}
\usage{
save_coverage(coverage, file)
}
\arguments{
\item{coverage}{A \code{coverage} object.}

\item{file}{A \code{character} text containg YAML file paths.}
}
\value{
\code{TRUE} if coverage is saved.
}
\description{
Save a \code{coverage} object to a YAML
}
