% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{stacks_functions}
\alias{stacks_functions}
\alias{stack_tiles}
\alias{stack_length}
\alias{stack_prune}
\alias{time_series}
\title{Stacks functions}
\usage{
stack_tiles(cube, bands = names(cube_bands(cube = cube)),
  start_dates = NULL, end_dates = NULL, which = NULL)

stack_length(stacks)

stack_prune(stacks, stack_length)

time_series(cube, longitude, latitude, bands = names(cube_bands(cube)),
  start_dates = NULL, end_dates = NULL, crs = cube_crs(cube))
}
\arguments{
\item{cube}{A \code{EOCubes_cube} data structure.}

\item{bands}{A \code{character} vector with band names to be retrieved.}

\item{start_dates}{A \code{Date} vector to filter layers by date.}

\item{end_dates}{A \code{Date} vector to filter layers by date.}

\item{which}{A \code{logical} or \code{integer} vector indicating which
tile to be stacked. If \code{NULL} (default) all tiles are stacked.}

\item{stacks}{A \code{EOCubes_stacks} object.}

\item{stack_length}{An \code{integer} indicating the length of stacks to
be filtered.}

\item{longitude}{A \code{numeric} value informing the longitute (X).}

\item{latitude}{A \code{numeric} value informing the latitude (Y).}

\item{crs}{An \code{integer} (EPSG code) or \code{character} (proj4)}
}
\value{
An \code{EOCubes_stacks} object.

A \code{list} of \code{integer}.

An \code{EOCubes_stacks} object.

An \code{EOCubes_timeseries} object.
}
\description{
These functions provides basic operations to work with stacks.
Stacks are a set of GeoTif reference locations organized by tile,
date interval, and bands.
}
\details{
\code{start_dates} and \code{end_dates} must have the same length.
These dates define intervals that are used to filter layers that constructs
stacks. The parameter \code{which} can be used to filter which tiles must
be stacked, hence which tiles must be fetched. Fetching tiles can be an
expensive task. You can filter tiles that intersects your area of interest
by providing \code{which} parameter.

\code{start_dates} and \code{end_dates} must have the same length.
These dates define intervals that are used to filter layers that constructs
the time series for each band and interval. \code{longitude} and
\code{latitude} parameters must be in \code{crs} projection. The default
value for \code{crs} is the same as the \code{cube} crs.
}
\section{Functions}{
\itemize{
\item \code{stack_tiles}: Returns a new \code{EOCubes_stacks} object that
represents stacks for each band and each date intervals.

\item \code{stack_length}: Returns a list of \code{integer} with the
lengths of each stack.

\item \code{stack_prune}: Returns an \code{EOCubes_stacks} object where
each stack has the same length given by \code{stack_length}.

\item \code{time_series}: Returns an \code{EOCubes_timeseries} object that
represents a set of band timeseries from some date intervals.
}}

\examples{
x <- remote("localhost")
cub1 <- cube("MOD13Q1/006", x)
tiles_index <- tiles_which(cub1, "h12v10")
stk1 <- stack_tiles(cub1, bands = c("evi", "ndvi"),
                    start_dates = "2014-01-01", end_dates = "2015-01-01",
                    which = tiles_index)
stack_length(stk1)
stack_prune(stk1, 23)

}
\seealso{
\code{\link{remote}}, \code{\link{cube}}, \code{\link{tiles_which}}
}
