% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{tiles_extents}
\alias{tiles_extents}
\title{Cubes management functions}
\usage{
tiles_extents(cube)
}
\arguments{
\item{cube}{A \code{EOCubes_cube} data structure.}
}
\value{
A \code{sfc} data structure with extents of all tiles in a cube.
}
\description{
Lists all registered tiles in a cube.
}
