% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{tiles_intersects}
\alias{tiles_intersects}
\title{Cubes management functions}
\usage{
tiles_intersects(cube, geom)
}
\arguments{
\item{cube}{A \code{EOCubes_cube} data structure.}

\item{geom}{A \code{sfc} or \code{sf} feature data from \code{sf} package to filter tiles.}
}
\value{
A \code{logical} vector.
}
\description{
Returns a \code{logical} vector indicating which tiles intersects \code{geom}.
}
\details{
If \code{cube} parameter is missing, the function will try to
retrieve the \code{EOCubes_cube} from the parent frame.
}
