% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{unpublish_coverage}
\alias{unpublish_coverage}
\title{Coverage management functions}
\usage{
unpublish_coverage(name, remote, ...)
}
\arguments{
\item{name}{A \code{character} text with the coverage name to be removed.}

\item{remote}{A \code{character} text with a valid remote server (with 'bucket' defined).}

\item{...}{Any extra arguments to be passed to \code{aws.s3} package to access the bucket.}
}
\value{
\code{TRUE} if coverage is published.
}
\description{
Removes a \code{coverage} object from a remote S3 server
This function uses \code{aws.s3} package to access the remote server.
You must provide appropriate access key to be able to write in the bucket.
}
