% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{version_functions}
\alias{version_functions}
\title{Version control functions}
\arguments{
\item{x}{A \code{list} data structure to be converted to repository object.}

\item{name}{A \code{character} text with repository name.}

\item{repository}{An \code{EOCubes_repository} object.}

\item{cube}{An \code{EOCubes_cube} object.}
}
\description{
These functions provides the basic operations over a cube
(\code{EOCubes_cube}) object.
}
\seealso{
\code{\link{repository}}
}
