% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{with_layers}
\alias{with_layers}
\title{Cubes management functions}
\usage{
with_layers(start_dates = NULL, end_dates = NULL, count = NULL)
}
\arguments{
\item{start_dates}{A \code{date} vector to filter layers by date.}

\item{end_dates}{A \code{date} vector to filter layers by date.}

\item{count}{An \code{integer} with the mandatory number of layers in each interval.}
}
\value{
A \code{list} parameters to select layers from stacks.
}
\description{
This is a helper function that construct a list of parameters
to be passed to \code{get_cube} function.
}
\details{
The \code{start_dates} and \code{end_dates} parameters values can
be a character vector containing dates in format \code{"\%Y-\%m-\%d"}.
Both parameters must have the same length.
}
