% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{with_stacks}
\alias{with_stacks}
\title{Cubes management functions}
\usage{
with_stacks(prefix = NULL)
}
\arguments{
\item{prefix}{A \code{character} vector containing stack name prefix(es)}
}
\value{
A \code{list} parameters to select stacks from tiles.
}
\description{
This is a helper function that construct a list of parameters
to be passed to \code{get_cube} function.
}
