% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{with_tiles}
\alias{with_tiles}
\title{Cubes management functions}
\usage{
with_tiles(prefix = NULL)
}
\arguments{
\item{prefix}{A \code{character} vector containing tile name prefix(es)}
}
\value{
A \code{list} parameters to select tiles from cubes.
}
\description{
This is a helper function that construct a list of parameters
to be passed to \code{get_cube} function.
}
