% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{default_remote}
\alias{default_remote}
\title{Remotes management functions}
\usage{
default_remote(default = NULL)
}
\value{
A \code{character} text with the name of default remote.
}
\description{
Fetches the default registered remote.
}
