% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.is_empty}
\alias{.is_empty}
\title{Internal functions}
\usage{
.is_empty(x)
}
\arguments{
\item{x}{Any object to be tested.}
}
\value{
\code{TRUE} if \code{x} is empty. \code{FALSE} otherwise.
}
\description{
Checks if a value is empty. The meaning of empty state depends on \code{x} class.
If \code{x} is a \code{data.frame}, the object is empty if it has zero columns or zero rows.
If \code{x} is a \code{character} vector, it is empty if \code{x} has length zero, has at least one \code{NA},
or \code{""} value. If \code{x} is a \code{numeric} vector, it is empty if it has length zero or has
at least one \code{NA} as its element. If \code{x} is a \code{list}, it is empty if it has length zero.
Otherwise, it is empty if it is \code{NULL} or has length zero.
}
\details{
#' @title Internal functions
#'
#' @description Open a brick file using \code{raster} package. If \code{file} is a url, it uses
#' GDAL driver \code{vsicurl}.
#'
#' @param file   A \code{character} text indicating a file path (local or in the web) to be opened.
#'
#' @return A \code{RasterBrick} object.
#'
.raster.open <- function(file) {

if (grepl("^https?://.+", file)) {

file = paste0("/vsicurl/", file)
    }

raster <- tryCatch(raster::brick(file, values = FALSE), error = function(e) NULL)

if (is.null(raster)) {

stop(sprintf("Error while opening raster '%s'.\nFile is unreachable.", file))
    }

return(brick)
}

#' @title Internal functions
#'
#' @description Test if all files in a bricks \code{data.frame} are reachable.
#' A progress bar will be showed if the number of files in bricks is greater or equal
#' the value of \code{getOption("progress_bar", 10)}.
#'
#' @param files   A bricks \code{data.frame}.
#'
#' @return \code{TRUE} if all bricks are reachable.
#'
.raster.touch <- function(files) {

progress_bar <- length(files) >= getOption(.progress_bar_option, 10)

if (progress_bar) {

pb <- utils::txtProgressBar(min = 0, max = length(files), style = 3)
    }

for (file in files) {

.raster.open(file)

if (progress_bar) {
            utils::setTxtProgressBar(pb = pb, value = utils::getTxtProgressBar(pb) + 1)
        }
    }

if (progress_bar) {
        close(pb)
    }

return(TRUE)
}
}
