% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.sublime_list}
\alias{.sublime_list}
\title{Internal functions}
\usage{
.sublime_list(x, conjunction = "and")
}
\arguments{
\item{x}{A \code{character} vector to be listed.}

\item{conjunction}{A \code{character} text to be used in output text.}
}
\value{
A \code{character} text.
}
\description{
List a human readable list of elements.
}
