% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{fetch_cube}
\alias{fetch_cube}
\alias{.fetch_cube}
\title{Cubes management functions}
\usage{
.fetch_cube(cube_ref)
}
\arguments{
\item{cube_ref}{A \code{character} text with cube name.}
}
\value{
A \code{EOCubes_cube} data structure.
}
\description{
Fetches a registered cube
}
