% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{list_cubes}
\alias{list_cubes}
\title{Remote functions}
\usage{
list_cubes(remote = default_remote(), prefix = NULL)
}
\arguments{
\item{remote}{An \code{EOCubes_remote} object.}

\item{prefix}{A \code{character} containing cube name prefix to be
filtered.}
}
\value{
An \code{EOCubes_cubelist} object or \code{NULL} if no cube
satisfies the filter criteria.
}
\description{
Lists all registered cubes in a remote.
}
\examples{
list_cubes()  # list cubes from default remote

}
