% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{remote_functions}
\alias{remote_functions}
\alias{list_remotes}
\alias{remote}
\alias{default_remote}
\alias{remote_name}
\title{Remotes functions}
\usage{
list_remotes(prefix = NULL)

remote(name)

default_remote(default = NULL)

remote_name(remote)
}
\arguments{
\item{prefix}{A \code{character} containing remote name prefix to be
filtered.}

\item{name}{A \code{character} text with remote name.}

\item{default}{A \code{character} text with the new default remote name.}

\item{remote}{An \code{EOCubes_remote} object.}
}
\value{
A \code{list} of remotes.

A \code{remote} data structure.

A \code{character} text with the name of default remote.

A \code{character}.
}
\description{
These functions provides a basic operations over remotes.
A remote is an entry point to a cube provider that can maintain many cubes.
}
\section{Functions}{
\itemize{
\item \code{list_remotes}: Lists all registered remotes.

\item \code{remote}: Fetches a registered remote.

\item \code{default_remote}: Fetches the default registered remote.

\item \code{remote_name}: Returns the name of a remote or a cube.
}}

\examples{
list_remotes()
x <- remote("localhost")
remote_name(x)  # shows 'localhost'

}
\seealso{
\code{\link{list_cubes}}
}
