% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{repository_functions}
\alias{repository_functions}
\alias{.default_localhost}
\alias{repository}
\alias{repository_name}
\alias{list_cubes}
\title{Repository functions}
\usage{
.default_localhost()

repository(name, caching = TRUE)

repository_name(repos)

list_cubes(repos = repository("localhost"), prefix = NULL)
}
\arguments{
\item{name}{A \code{character} text with repository name.}

\item{caching}{A \code{logical} value indicating wether to use cache system.}

\item{repos}{An \code{EOCubes_repository} object.}

\item{prefix}{A \code{character} containing repository/cube entry name
prefix to be filtered.}

\item{x}{A \code{list} data structure to be converted
to \code{EOCubes_repository} object.}
}
\value{
A \code{EOCubes_repository} object.

A \code{repository} data structure.

A \code{character}.

An \code{EOCubes_cubelist} object or \code{NULL} if no cube
satisfies the filter criteria.
}
\description{
These functions provides a basic operations over repositories.
A repository is an entry point to a cube provider that can maintain many cubes.
}
\details{
The function \code{repository} fetches the repository registered on a given
entry \code{name}.

The function \code{repository_name} show the entry name of the repository
list from which the \code{EOCubes_repository} object have been fetched.

The function \code{list_cubes} lists all registered cubes entries
in a repository. Use \code{prefix} parameter to filter the entries by name.
}
\section{Functions}{
\itemize{
\item \code{.default_localhost}: Return the default localhost repository.

\item \code{repository}: Fetches a registered repository.

\item \code{repository_name}: Returns the name of a repository or a cube.

\item \code{list_cubes}: Lists all registered cubes in a repository.
}}

\examples{
list_repositories()
x <- repository("localhost")
repository_name(x)   # shows 'localhost'
list_cubes(x)   # list cubes in 'localhost'

}
\seealso{
\code{\link{list_cubes}}
}
