% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{stacks_functions}
\alias{stacks_functions}
\alias{stacks}
\alias{stacks_length}
\alias{stacks_prune}
\title{Stacks functions}
\usage{
stacks(cube, start_dates, end_dates, bands = names(cube_bands(cube =
  cube)))

stacks_length(stacks)

stacks_prune(stacks, stack_length)
}
\arguments{
\item{cube}{A \code{EOCubes_cube} data structure.}

\item{start_dates}{A \code{Date} vector to filter layers by date.}

\item{end_dates}{A \code{Date} vector to filter layers by date.}

\item{bands}{A \code{character} vector with band names to be retrieved.}

\item{stacks}{A \code{EOCubes_stacks} object.}

\item{stack_length}{An \code{integer} indicating the length of stacks to
be filtered.}
}
\value{
An \code{EOCubes_stacks} object.

A \code{list} of \code{integer}.

An \code{EOCubes_stacks} object.
}
\description{
These functions provides basic operations to work with stacks.
Stacks are a set of GeoTif reference locations organized by tile,
date interval, and bands.
}
\details{
\code{start_dates} and \code{end_dates} must have the same length.
These dates define intervals that are used to filter layers that constructs
stacks.
}
\section{Functions}{
\itemize{
\item \code{stacks}: Returns a new \code{EOCubes_stacks} object that
represents stacks for each band and each date intervals.

\item \code{stacks_length}: Returns a list of \code{integer} with the
lengths of each stack.

\item \code{stacks_prune}: Returns an \code{EOCubes_stacks} object where
each stack has the same length given by \code{stack_length}.
}}

\examples{
x <- remote("localhost")
cub1 <- cube("MOD13Q1/006", x)
cub2 <- filter_tiles(cub1, "h13v10")
cub3 <- fetch_tiles(cub2)
stk1 <- stacks(cub3, "2010-09-01", "2011-09-01", c("evi", "ndvi"))
stacks_length(stk1)
stk2 <- stacks_prune(stk1, 23)

}
\seealso{
\code{\link{remote}}, \code{\link{cube}}
}
